+++ src/common/schedulers/slurm_commands.py
@@ -235,7 +235,7 @@ def update_all_partitions(state, reset_node_addrs_hostname):
                 logging.info(f"Setting partition {part.name} state from {part.state} to {state}")
                 if reset_node_addrs_hostname:
                     logging.info(f"Resetting partition nodes {part.nodes}")
-                    set_nodes_down_and_power_save(part.nodes, reason="stopping cluster")
+                    set_nodes_power_down(part.nodenames, reason="stopping cluster")
                 partition_to_update.append(part.name)
         succeeded_partitions = update_partitions(partition_to_update, state)
         return succeeded_partitions == partition_to_update
@@ -293,7 +293,7 @@ def set_nodes_drain(nodes, reason):
 
 def set_nodes_power_down(nodes, reason=None):
     """Place slurm node into power_down state and reset nodeaddr/nodehostname."""
-    reset_nodes(nodes=nodes, state="power_down", reason=reason, raise_on_error=True)
+    reset_nodes(nodes=nodes, state="power_down_force", reason=reason, raise_on_error=True)
 
 
 def reset_nodes(nodes, state=None, reason=None, raise_on_error=False):
@@ -319,18 +319,6 @@ def set_nodes_idle(nodes, reason=None, reset_node_addrs_hostname=False):
         reset_nodes(nodes, state="resume", reason=reason, raise_on_error=False)
     else:
         update_nodes(nodes=nodes, state="resume", reason=reason, raise_on_error=False)
-
-
-@retry(stop_max_attempt_number=3, wait_fixed=1500)
-def set_nodes_down_and_power_save(node_list, reason):
-    """
-    Set slurm nodes into down -> power_down.
-
-    This is the standard failure recovery procedure to reset a CLOUD node.
-    """
-    set_nodes_down(node_list, reason=reason)
-    set_nodes_power_down(node_list, reason=reason)
-
 
 def get_nodes_info(nodes="", command_timeout=DEFAULT_GET_INFO_COMMAND_TIMEOUT):
     """