+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This file is used to list changes made in each version of the aws-parallelcluste
 2.11.8
 ------
 **CHANGES**
-- Upgrade Slurm to 21.08
+- Update Slurm plugin to support version 21.08.8-2
 
 2.11.7
 ------